# optimize MergingIterator with sorted children