In current implemetation, children of MergingIterator are
unordered. Thus FindSmallest() and FindLargest() have to scan all
children to find a proper current_ and introduce n_ comparation. When
application iterating with iterator, there are many duplicated
comparations.

In this patch, we make children ordered, so that FindSmallest() and
FindLargest() don't need to scan all children but just pick the first
valid from the begin and the end respectively. When current_ is
changed by Next() or Prev(), we would move current_ forward or
backward to ensure the order of children.

This method improves the performance about 23% when immediatedly
iterating the db that are randomly inserted with a huge number of
records.

Performance without patch

LevelDB:    version 1.22
Date:       Tue Sep 10 23:38:20 2019
CPU:        24 * Intel(R) Xeon(R) CPU E5-2620 v2 @ 2.10GHz
CPUCache:   15360 KB
Keys:       16 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1048576
RawSize:    116.0 MB (estimated)
FileSize:   66.0 MB (estimated)
------------------------------------------------
fillrandom   :       4.537 micros/op;   24.4 MB/s
readseq      :       0.569 micros/op;  194.5 MB/s
compact      : 1318885.000 micros/op;
readseq      :       0.260 micros/op;  426.0 MB/s

Performance with this patch

LevelDB:    version 1.22
Date:       Tue Sep 10 23:37:36 2019
CPU:        24 * Intel(R) Xeon(R) CPU E5-2620 v2 @ 2.10GHz
CPUCache:   15360 KB
Keys:       16 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1048576
RawSize:    116.0 MB (estimated)
FileSize:   66.0 MB (estimated)
------------------------------------------------
fillrandom   :       4.446 micros/op;   24.9 MB/s
readseq      :       0.460 micros/op;  240.6 MB/s
compact      : 1570193.000 micros/op;
readseq      :       0.259 micros/op;  426.8 MB/s

Signed-off-by: Kyle Zhang <kyle@smartx.com>