@kylezh 2019-09-10
@pwnall @cmumford Could you please retrigger the CI tests? It seems the failure is caused by a network error on CI server.

```
==> Installing gdal
==> Downloading https://homebrew.bintray.com/bottles/gdal-2.4.2_2.mojave.bottle.
curl: (6) Could not resolve host: homebrew.bintray.com
Error: Failed to download resource "gdal"
Download failed: https://homebrew.bintray.com/bottles/gdal-2.4.2_2.mojave.bottle.tar.gz
Warning: Bottle installation failed: building from source.
==> Downloading https://download.osgeo.org/gdal/2.4.2/gdal-2.4.2.tar.xz
######################################################################## 100.0%
==> ./configure --prefix=/usr/local/Cellar/gdal/2.4.2_2 --mandir=/usr/local/Cell
==> make
No output has been received in the last 10m0s, this potentially indicates a stalled build or something wrong with the build itself.
Check the details on how to adjust your build configuration on: https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received
The build has been terminated
```